feat: :sparkles: I implemented the login form with client-side validation.
I added the `SigninPage` component with a form for user login.
It includes input fields for email and password, a loading spinner, and toast notifications for success and error messages.
I used the `next-auth` `signIn` method to handle login without redirection.
A link to the signup page was also added for users without an account.